[github_actions] Use env variable for the project name
when copying over source files for the RPM build. This is
consistent with our other workflows and fixes a current issue
where all source files (from every RPM in this project) is
incorrectly trying to be copied over. Previously there was a
$DIRECTORY variable being used that was empty so the 'cp' command
would instead grab all files from the root directory of the git
project.